G31 Z-Axis Crash Fix
The firmware that ships with the He3D DLT-600 has a bug in it that causes the print head to crash into the build plate. As part of the G31 auto-leveling command, the last step executed is to move X and Y to zero, and Z to 5. Depending upon the height of the print bed (or build plate), 5 can be below the surface. With this fix in place, the G31 command will cause the z-probing to occur. After all of the probes, the print head will move to the center of the plate somewhere shy of 30mm high (Z axis) before it begins printing.